requirements.txt

pandas
openpyxl
PyGithub
